requirements.txt
requests==2.32.3
pandas==2.2.2
duckdb==1.0.0
pyyaml==6.0.2
lxml==6.0.2
matplotlib
yfinance